Rename AddCustomerCommand test cases
