--- src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -1,311 +1,0 @@
-package seedu.address.logic.commands;
-
-import static java.util.Objects.requireNonNull;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static seedu.address.testutil.Assert.assertThrows;
-
-import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Set;
-import java.util.function.Predicate;
-
-import org.junit.jupiter.api.Test;
-
-import javafx.collections.ObservableList;
-import seedu.address.commons.core.GuiSettings;
-import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.AddressBook;
-import seedu.address.model.Model;
-import seedu.address.model.ReadOnlyAddressBook;
-import seedu.address.model.ReadOnlyUserPrefs;
-import seedu.address.model.UserPrefs;
-import seedu.address.model.cheese.Cheese;
-import seedu.address.model.cheese.CheeseId;
-import seedu.address.model.cheese.CheeseType;
-import seedu.address.model.customer.Customer;
-import seedu.address.model.customer.Phone;
-import seedu.address.model.order.Order;
-import seedu.address.model.order.Quantity;
-import seedu.address.testutil.CustomerBuilder;
-
-public class AddCommandTest {
-
-    @Test
-    public void constructor_nullCustomer_throwsNullPointerException() {
-        assertThrows(NullPointerException.class, () -> new AddCustomerCommand(null));
-    }
-
-    @Test
-    public void execute_customerAcceptedByModel_addSuccessful() throws Exception {
-        ModelStubAcceptingCustomerAdded modelStub = new ModelStubAcceptingCustomerAdded();
-        Customer validCustomer = new CustomerBuilder().build();
-
-        CommandResult commandResult = new AddCustomerCommand(validCustomer).execute(modelStub);
-
-        assertEquals(String.format(AddCustomerCommand.MESSAGE_SUCCESS, validCustomer),
-                commandResult.getFeedbackToUser());
-        assertEquals(Arrays.asList(validCustomer), modelStub.customersAdded);
-        assertTrue(modelStub.getGuiSettings().isShowCustomerListPanel());
-    }
-
-    @Test
-    public void execute_duplicateCustomer_throwsCommandException() {
-        Customer validCustomer = new CustomerBuilder().build();
-        AddCustomerCommand addCommand = new AddCustomerCommand(validCustomer);
-        ModelStub modelStub = new ModelStubWithCustomer(validCustomer);
-
-        assertThrows(CommandException.class, AddCustomerCommand.MESSAGE_DUPLICATE_CUSTOMER, () -> addCommand
-                .execute(modelStub));
-    }
-
-    @Test
-    public void equals() {
-        Customer alice = new CustomerBuilder().withName("Alice").build();
-        Customer bob = new CustomerBuilder().withName("Bob").build();
-        AddCustomerCommand addAliceCommand = new AddCustomerCommand(alice);
-        AddCustomerCommand addBobCommand = new AddCustomerCommand(bob);
-
-        // same object -> returns true
-        assertTrue(addAliceCommand.equals(addAliceCommand));
-
-        // same values -> returns true
-        AddCustomerCommand addAliceCommandCopy = new AddCustomerCommand(alice);
-        assertTrue(addAliceCommand.equals(addAliceCommandCopy));
-
-        // different types -> returns false
-        assertFalse(addAliceCommand.equals(1));
-
-        // null -> returns false
-        assertFalse(addAliceCommand.equals(null));
-
-        // different customer -> returns false
-        assertFalse(addAliceCommand.equals(addBobCommand));
-    }
-
-    /**
-     * A default model stub that have all of the methods failing.
-     */
-    private class ModelStub implements Model {
-        @Override
-        public void setUserPrefs(ReadOnlyUserPrefs userPrefs) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyUserPrefs getUserPrefs() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setGuiSettings(GuiSettings guiSettings) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Path getAddressBookFilePath() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBookFilePath(Path addressBookFilePath) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addCustomer(Customer customer) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setAddressBook(ReadOnlyAddressBook newData) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasCustomer(Customer customer) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteCustomer(Customer target) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setCustomer(Customer target, Customer editedCustomer) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasOrder(Order order) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteOrder(Order order) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addOrder(Order order) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setOrder(Order target, Order editedOrder) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasCheese(Cheese cheese) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void deleteCheese(Cheese cheese) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void addCheese(Cheese cheese) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setCheese(Cheese target, Cheese editedCheese) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Set<CheeseId> getUnassignedCheeses(CheeseType cheesetype, Quantity quantity) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateCheesesStatus(Set<CheeseId> cheesesAssigned) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Customer> getFilteredCustomerList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Order> getFilteredOrderList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public ObservableList<Cheese> getFilteredCheeseList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredCustomerList(Predicate<Customer> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredOrderList(Predicate<Order> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void updateFilteredCheeseList(Predicate<Cheese> predicate) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPanelToCustomerList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPanelToCheeseList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public void setPanelToOrderList() {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public boolean hasCustomerWithPhone(Phone phone) {
-            throw new AssertionError("This method should not be called.");
-        }
-
-        @Override
-        public Customer getCustomerWithPhone(Phone phone) {
-            throw new AssertionError("This method should not be called.");
-        }
-    }
-
-    /**
-     * A Model stub that contains a single customer.
-     */
-    private class ModelStubWithCustomer extends ModelStub {
-        private final Customer customer;
-
-        ModelStubWithCustomer(Customer customer) {
-            requireNonNull(customer);
-            this.customer = customer;
-        }
-
-        @Override
-        public boolean hasCustomer(Customer customer) {
-            requireNonNull(customer);
-            return this.customer.isSameCustomer(customer);
-        }
-    }
-
-    /**
-     * A Model stub that always accept the customer being added.
-     */
-    private class ModelStubAcceptingCustomerAdded extends ModelStub {
-        final ArrayList<Customer> customersAdded = new ArrayList<>();
-        private final ReadOnlyUserPrefs userPrefs = new UserPrefs();
-
-        @Override
-        public boolean hasCustomer(Customer customer) {
-            requireNonNull(customer);
-            return customersAdded.stream().anyMatch(customer::isSameCustomer);
-        }
-
-        @Override
-        public void addCustomer(Customer customer) {
-            requireNonNull(customer);
-            customersAdded.add(customer);
-        }
-
-        @Override
-        public GuiSettings getGuiSettings() {
-            return userPrefs.getGuiSettings();
-        }
-
-        @Override
-        public void setPanelToCustomerList() {
-            getGuiSettings().setPanelToCustomerList();
-        }
-
-        @Override
-        public ReadOnlyAddressBook getAddressBook() {
-            return new AddressBook();
-        }
-    }
-
-}
